Fix some issues after moving the model
