+++ readthedocs/api/v2/utils.py
@@ -15,9 +15,9 @@ from readthedocs.builds.constants import NON_REPOSITORY_VERSIONS
 from readthedocs.builds.constants import STABLE
 from readthedocs.builds.constants import STABLE_VERBOSE_NAME
 from readthedocs.builds.constants import TAG
-from readthedocs.builds.models import AutomationRule
 from readthedocs.builds.models import Version
 from readthedocs.core.utils.db import delete_in_batches
+from readthedocs.projects.models import AutomationRule
 
 
 log = structlog.get_logger(__name__)
